algorithms/sbd/pyproject.toml
[build-system]
requires = ["setuptools>=64"]
build-backend = "setuptools.build_meta"

[project]
name = "sbd"
version = "0.1.0"
description = "Prefect workflow of SBD experiment from arXiv:2511.00224."
authors = [
    { name = "Yuto Morohoshi", email = "mrhsyut@ibm.com" }
]
maintainers = [
    { name = "Yuto Morohoshi", email = "mrhsyut@ibm.com" }
]
requires-python = ">=3.12"
dependencies = [
    "ffsim==0.0.57",
    "qiskit-addon-sqd==0.11.0",
    "prefect-qiskit==0.2.0",
    "prefect[aws,ray]==3.4.24",
    "pydantic_numpy",
    "prefect-sbd",
    "prefect-miyabi",
    "qcsc_workflow_utility",
]

[tool.uv.sources]
prefect-sbd = { workspace = true }
prefect-miyabi = { workspace = true }
qcsc_workflow_utility = { workspace = true }

[tool.setuptools]
packages = ["sbd"]
include-package-data = true
zip-safe = false

[project.scripts]
sbd-deploy = "sbd.main:deploy"
